ascon: expose permutation
Provides raw access to the Ascon permutation, allowing it to be used for
more things than just AEAD.
