+++ ascon/benches/bench.rs
@@ -1,9 +1,8 @@
 #![feature(test)]
 
-extern crate ascon;
 extern crate test;
 
-use ascon::{aead_decrypt, aead_encrypt};
+use ascon::aead;
 use test::Bencher;
 
 #[bench]
@@ -14,7 +13,7 @@ fn ascon_encrypt_bench(b: &mut Bencher) {
     let message = [99; 1025];
 
     b.bytes = message.len() as u64;
-    b.iter(|| aead_encrypt(&key, &iv, &message, &aad));
+    b.iter(|| aead::encrypt(&key, &iv, &message, &aad));
 }
 
 #[bench]
@@ -23,8 +22,8 @@ fn ascon_decrypt_bench(b: &mut Bencher) {
     let iv = [8; 16];
     let aad = [3; 16];
     let message = [99; 1025];
-    let (ciphertext, tag) = aead_encrypt(&key, &iv, &message, &aad);
+    let (ciphertext, tag) = aead::encrypt(&key, &iv, &message, &aad);
 
     b.bytes = message.len() as u64;
-    b.iter(|| aead_decrypt(&key, &iv, &ciphertext, &aad, &tag));
+    b.iter(|| aead::decrypt(&key, &iv, &ciphertext, &aad, &tag));
 }
